GeoreferencingDialog: Fix compiler warning
The Q_STATIC_ASSERT at the top of the file caused a compiler warning:
comparison of integer expressions of different signedness:
‘unsigned int’ and ‘int’

+++ src/gui/georeferencing_dialog.cpp
@@ -88,7 +88,7 @@
 
 namespace OpenOrienteering {
 
-Q_STATIC_ASSERT(Georeferencing::declinationPrecision() == Util::InputProperties<Util::RotationalDegrees>::decimals());
+Q_STATIC_ASSERT(Georeferencing::declinationPrecision() == static_cast<unsigned int>(Util::InputProperties<Util::RotationalDegrees>::decimals()));
 
 
 namespace  {
